- renamed apply-for-benefits directory to benefit-application-drafts - renamed SavedApplications component to BenefitApplicationDrafts
