+++ src/applications/personalization/dashboard/components/benefit-application-drafts/BenefitOfInterest.jsx
