+++ src/applications/personalization/dashboard/tests/components/apply-for-benefits/ApplicationInProgress.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { Toggler } from '~/platform/utilities/feature-toggles';
 import { renderWithStoreAndRouter } from '~/platform/testing/unit/react-testing-library-helpers';
 
-import ApplicationInProgress from '../../../components/apply-for-benefits/ApplicationInProgress';
+import ApplicationInProgress from '../../../components/benefit-application-drafts/ApplicationInProgress';
 
 describe('ApplicationInProgress component', () => {
   // delete instances of this toggle and use of renderWithStoreAndRouter when #68314 is launched
