+++ src/applications/personalization/dashboard/tests/components/apply-for-benefits/BenefitApplicationDrafts.unit.spec.jsx
@@ -3,12 +3,12 @@ import { expect } from 'chai';
 import { renderInReduxProvider } from '~/platform/testing/unit/react-testing-library-helpers';
 import reducers from '~/applications/personalization/dashboard/reducers';
 
-import SavedApplications from '../../../components/apply-for-benefits/SavedApplications';
+import BenefitApplicationDrafts from '../../../components/benefit-application-drafts/BenefitApplicationDrafts';
 
-describe('SavedApplications component', () => {
+describe('BenefitApplicationDrafts component', () => {
   it('renders correctly', () => {
     const initialState = {};
-    const view = renderInReduxProvider(<SavedApplications />, {
+    const view = renderInReduxProvider(<BenefitApplicationDrafts />, {
       initialState,
       reducers,
     });
